# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Web+ API
  description: |
    Web+ is a comprehensive platform for managing and interacting with large language models.
    This API provides endpoints for authentication, model management, chat completions,
    file handling, and pipeline automation.
  version: 0.2.0
  contact:
    name: Web+ Support
    email: support@webplus.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8002/api
    description: Local development server
  - url: https://api.webplus.ai
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /pipelines:
    get:
      tags: [Pipelines]
      summary: List pipelines
      operationId: listPipelines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of pipelines
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pipeline'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Pipelines]
      summary: Create a pipeline
      operationId: createPipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineCreateRequest'
      responses:
        '201':
          description: Pipeline created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Pipeline:
      type: object
      required: [id, name, description, steps]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/PipelineStep'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    PipelineStep:
      type: object
      required: [id, name, type, order]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [llm_prompt, file_processor, code_executor, conditional, loop]
        order:
          type: integer
        config:
          type: object
    PipelineCreateRequest:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
            required: [name, type, config]
            properties:
              name:
                type: string
              type:
                type: string
              config:
                type: object
...
    Pagination:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...